+++ Gra-Pipe/Gra-Pipe/Scene.cpp
@@ -1,7 +1,7 @@
 #include "Scene.h"
 
 int Scene::getHighlightedID() {
-	for (int i = 0; i < this->buttons.size(); i++) {
+	for (size_t i = 0; i < this->buttons.size(); i++) {
 		if (this->buttons.at(i)->isSeleced) {
 			return i;
 		}
